Validação no front

+++ resources/views/produto/index.blade.php
@@ -34,12 +34,14 @@
                 {{$produto->titulo}}
             </a>
         @endif
-        {{Form::open(['route' => ['produtos.destroy', $produto->id],
-            'method'=>'DELETE'
-        ])}}
-        <a href="{{url('produtos/'.$produto->id.'/edit')}}" class="btn btn-default">Editar</a>
-        {{Form::submit('Excluir', ['class' => 'btn btn-default'])}}
-        {{Form::close()}}
+        @if(Auth::check())
+            {{Form::open(['route' => ['produtos.destroy', $produto->id],
+                'method'=>'DELETE'
+            ])}}
+            <a href="{{url('produtos/'.$produto->id.'/edit')}}" class="btn btn-default">Editar</a>
+            {{Form::submit('Excluir', ['class' => 'btn btn-default'])}}
+            {{Form::close()}}
+        @endif
     </div>
 
 @endforeach
